Feature/add php stan

+++ uninstall.php
@@ -16,4 +16,9 @@ if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 	exit;
 }
 
+// do nothing if PHP-version is not 8.0 or newer.
+if ( PHP_VERSION_ID < 80000 ) { // @phpstan-ignore smaller.alwaysFalse
+	return;
+}
+
 delete_option( 'audio-on-every-block' );
